Copy over all the local files but without the .ply file.
This loses a bunch of my git history, but I can't back up my current local repo to github because there is too large of a file in it. So I've checked out the repo again and removed the large file. You will have to download your own .ply file and set it to src/assets/example.ply to test out the code.